Moved UDP parser to parsers
